rename lambda_func.py to f.py and handler() to f()
