--- test-registry/hello/lambda_func.py
@@ -1,2 +1,0 @@
-def handler(event):
-    return 'hello'
