+++ test-registry/install2/f.py
@@ -3,5 +3,5 @@ import urllib3
 
 # ol-install: requests,certifi,chardet,idna,urllib3
 
-def handler(event):
+def f(event):
     return 'imported'
