+++ test-registry/numpy15/f.py
@@ -1,5 +1,5 @@
 # ol-install: numpy==1.15
 import numpy
 
-def handler(event):
+def f(event):
     return {'result': int(numpy.array(event).sum()), 'version': numpy.__version__}
