+++ test-registry/pandas15/f.py
@@ -1,0 +1,7 @@
+# ol-install: numpy==1.15,pandas
+import numpy
+import pandas
+
+def f(event):
+    df = pandas.DataFrame(event)
+    return {'result': int(df.values.sum()), 'version': numpy.__version__}
